databaseCon files moved to sub directory called "data"
