+++ data/databasecon.h
@@ -6,11 +6,18 @@
 #include <QSqlDatabase>
 #include <QSqlQuery>
 #include <QVector>
+#include <QtDebug>
+#include <QSqlError>
+#include <QtXml>
+#include <QDir>
+#include <QFile>
+#include "data/xmlmanipulation.h"
 
 class databaseCon
 {
 public:
     databaseCon();
+    static void initDB();
     QSqlQuery* excequte(QString cmdstr);
 
 private:
